Adds Kubernetes with Syncope 2.0.6
This is the basic install for Syncope into kubernetes.   It is very basic but it is a good starting point for any production insteallation.   The passwords are not secure and it is not configured for an HA environment.   It works on my minikube and should be a good starting point.

I changed all the repos as I was pointing to a private repo.   I moved them over to a public dockerhub and put in generic containers that could be used by anyone.

There are some small improvments that could be done, better docs on setting up K8s, cleaning up the debs, etc.   But it should be good enough :)